requirements.txt
fastapi~=0.116.1
SQLAlchemy~=2.0.43
uvicorn~=0.35.0
pydantic~=2.11.7
PyJWT~=2.10.1
cryptography~=45.0.6
passlib~=1.7.4
pytest~=8.4.1
pytest-asyncio
httpx~=0.28.1
a2wsgi~=1.10.10
SQLAlchemy-Utils~=0.41.2
python-dotenv~=1.1.1